Fixes issue # 2231 Rename the provider files and separate classes
1. Create new _defaultmethodprovider.py
2. Create new file _providerdispatcher.py
3. Modify imports, etc. to use the new files.

I kept the name defaultinstanceprovider because I am uncomfortable
creating separate providers in the future for instancewrite and
instanceread.  DISCUSSION.
